New sound system (SDL + SteamAudio)
Fix SDK and impl phonon reverb

SDL3 update (no nuget bitches)

Implement voice chat

Added steam audio disable define

Implement switch audio device

Implement `EAX` values change

Fix attenuation and reverb

Memory opt

Fix rain steps volume

hype fixes (leaks, attenuation, perf, doors)

Restore total size and total time in `Sound Editor`

Синхра с фиксами

Add backpack animator

Clang: cleanup, fixes, dead code

ClangCI: We can start the game

+++ src/xrGame/VoiceChat.h
@@ -47,7 +47,7 @@ private:
 	void CheckAndClearPlayers(SOUND_PLAYERS& players);
 
 private:
-	u8 m_buffer[1024];
+	float m_buffer[1024];
 
 	ISoundVoiceChat* m_pSoundVoiceChat = nullptr;
 	ISoundRecorder* m_pRecorder = nullptr;
